chore: Moved generate locations from StartGameHandler to DungeonMaster
